working further abstractions for client
