+++ cmd/gaia/cmd/gaiacli/main.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"net/http"
 	"os"
-	"path"
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/keys"
@@ -40,10 +39,8 @@ import (
 
 	"github.com/rakyll/statik/fs"
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
 
 	amino "github.com/tendermint/go-amino"
-	"github.com/tendermint/tendermint/libs/cli"
 
 	_ "github.com/cosmos/cosmos-sdk/client/lcd/statik"
 )
@@ -66,6 +63,7 @@ func main() {
 	// the below functions and eliminate global vars, like we do
 	// with the cdc
 
+	// XXX abstract this
 	// Module clients hold cli commnads (tx,query) and lcd routes
 	// TODO: Make the lcd command take a list of ModuleClient
 	mc := []sdk.ModuleClient{
@@ -82,18 +80,12 @@ func main() {
 		Short: "Command line interface for interacting with gaiad",
 	}
 
-	// Add --chain-id to persistent flags and mark it required
-	rootCmd.PersistentFlags().String(client.FlagChainID, "", "Chain ID of tendermint node")
-	rootCmd.PersistentPreRunE = func(_ *cobra.Command, _ []string) error {
-		return initConfig(rootCmd)
-	}
-
 	// Construct Root Command
 	rootCmd.AddCommand(
 		rpc.StatusCommand(),
 		client.ConfigCmd(app.DefaultCLIHome),
-		queryCmd(cdc, mc),
-		txCmd(cdc, mc),
+		queryCmd(cdc, app.BasicGaiaApp),
+		txCmd(cdc, app.BasicGaiaApp),
 		client.LineBreak,
 		lcd.ServeCommand(cdc, registerRoutes),
 		client.LineBreak,
@@ -104,7 +96,7 @@ func main() {
 	)
 
 	// Add flags and prefix all env exposed with GA
-	executor := cli.PrepareMainCmd(rootCmd, "GA", app.DefaultCLIHome)
+	executor := client.PrepareMainCmd(rootCmd, "GA", app.DefaultCLIHome)
 
 	err := executor.Execute()
 	if err != nil {
@@ -113,13 +105,12 @@ func main() {
 	}
 }
 
-func queryCmd(cdc *amino.Codec, mc []sdk.ModuleClient) *cobra.Command {
+func queryCmd(cdc *amino.Codec, mbm sdk.ModuleBasicManager) *cobra.Command {
 	queryCmd := &cobra.Command{
 		Use:     "query",
 		Aliases: []string{"q"},
 		Short:   "Querying subcommands",
 	}
-
 	queryCmd.AddCommand(
 		rpc.ValidatorCommand(cdc),
 		rpc.BlockCommand(),
@@ -128,23 +119,15 @@ func queryCmd(cdc *amino.Codec, mc []sdk.ModuleClient) *cobra.Command {
 		client.LineBreak,
 		authcmd.GetAccountCmd(at.StoreKey, cdc),
 	)
-
-	for _, m := range mc {
-		mQueryCmd := m.GetQueryCmd()
-		if mQueryCmd != nil {
-			queryCmd.AddCommand(mQueryCmd)
-		}
-	}
-
+	mbm.AddQueryCommands(txCmd)
 	return queryCmd
 }
 
-func txCmd(cdc *amino.Codec, mc []sdk.ModuleClient) *cobra.Command {
+func txCmd(cdc *amino.Codec, mbm sdk.ModuleBasicManager) *cobra.Command {
 	txCmd := &cobra.Command{
 		Use:   "tx",
 		Short: "Transactions subcommands",
 	}
-
 	txCmd.AddCommand(
 		bankcmd.SendTxCmd(cdc),
 		client.LineBreak,
@@ -154,21 +137,20 @@ func txCmd(cdc *amino.Codec, mc []sdk.ModuleClient) *cobra.Command {
 		tx.GetEncodeCommand(cdc),
 		client.LineBreak,
 	)
-
-	for _, m := range mc {
-		txCmd.AddCommand(m.GetTxCmd())
-	}
-
+	mbm.AddTxCommands(txCmd)
 	return txCmd
 }
 
 // registerRoutes registers the routes from the different modules for the LCD.
 // NOTE: details on the routes added for each module are in the module documentation
 // NOTE: If making updates here you also need to update the test helper in client/lcd/test_helper.go
-func registerRoutes(rs *lcd.RestServer) {
+func registerRoutes(rs *lcd.RestServer, mbm sdk.ModuleBasicManager) {
 	registerSwaggerUI(rs)
 	rpc.RegisterRoutes(rs.CliCtx, rs.Mux)
 	tx.RegisterRoutes(rs.CliCtx, rs.Mux, rs.Cdc)
+	mbm.RegisterRESTRoutes(rs.CliCtx, rs.Mux, rs.Cdc, rs.KeyBase)
+
+	// XXX Abstract this
 	auth.RegisterRoutes(rs.CliCtx, rs.Mux, rs.Cdc, at.StoreKey)
 	bank.RegisterRoutes(rs.CliCtx, rs.Mux, rs.Cdc, rs.KeyBase)
 	dist.RegisterRoutes(rs.CliCtx, rs.Mux, rs.Cdc, distcmd.StoreKey)
@@ -185,27 +167,4 @@ func registerSwaggerUI(rs *lcd.RestServer) {
 	}
 	staticServer := http.FileServer(statikFS)
 	rs.Mux.PathPrefix("/swagger-ui/").Handler(http.StripPrefix("/swagger-ui/", staticServer))
-}
-
-func initConfig(cmd *cobra.Command) error {
-	home, err := cmd.PersistentFlags().GetString(cli.HomeFlag)
-	if err != nil {
-		return err
-	}
-
-	cfgFile := path.Join(home, "config", "config.toml")
-	if _, err := os.Stat(cfgFile); err == nil {
-		viper.SetConfigFile(cfgFile)
-
-		if err := viper.ReadInConfig(); err != nil {
-			return err
-		}
-	}
-	if err := viper.BindPFlag(client.FlagChainID, cmd.PersistentFlags().Lookup(client.FlagChainID)); err != nil {
-		return err
-	}
-	if err := viper.BindPFlag(cli.EncodingFlag, cmd.PersistentFlags().Lookup(cli.EncodingFlag)); err != nil {
-		return err
-	}
-	return viper.BindPFlag(cli.OutputFlag, cmd.PersistentFlags().Lookup(cli.OutputFlag))
 }
