+++ x/auth/types/account_test.go
@@ -1,4 +1,4 @@
-package auth
+package types
 
 import (
 	"testing"
