+++ x/auth/types/codec.go
@@ -1,4 +1,4 @@
-package auth
+package types
 
 import (
 	"github.com/cosmos/cosmos-sdk/codec"
