+++ x/auth/types/feekeeper_test.go
@@ -1,4 +1,4 @@
-package auth
+package types
 
 import (
 	"testing"
