+++ x/auth/types/stdtx_test.go
@@ -1,4 +1,4 @@
-package auth
+package types
 
 import (
 	"fmt"
