+++ x/bank/client/rest/sendtx.go
@@ -12,7 +12,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/rest"
 
-	"github.com/cosmos/cosmos-sdk/x/bank"
+	"github.com/cosmos/cosmos-sdk/x/bank/types"
 )
 
 // RegisterRoutes - Central function to define routes that get registered by the main application
@@ -24,12 +24,6 @@ func RegisterRoutes(cliCtx context.CLIContext, r *mux.Router, cdc *codec.Codec, 
 type SendReq struct {
 	BaseReq rest.BaseReq `json:"base_req"`
 	Amount  sdk.Coins    `json:"amount"`
-}
-
-var moduleCdc = codec.New()
-
-func init() {
-	bank.RegisterCodec(moduleCdc)
 }
 
 // SendRequestHandlerFn - http request handler to send coins to a address.
@@ -60,7 +54,7 @@ func SendRequestHandlerFn(cdc *codec.Codec, kb keys.Keybase, cliCtx context.CLIC
 			return
 		}
 
-		msg := bank.NewMsgSend(fromAddr, toAddr, req.Amount)
+		msg := types.NewMsgSend(fromAddr, toAddr, req.Amount)
 		clientrest.WriteGenerateStdTxResponse(w, cdc, cliCtx, req.BaseReq, []sdk.Msg{msg})
 	}
 }
