+++ x/bank/module.go
@@ -3,9 +3,12 @@ package bank
 import (
 	"encoding/json"
 
+	"github.com/cosmos/cosmos-sdk/client/context"
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/auth"
+	"github.com/cosmos/cosmos-sdk/x/gov/client/rest"
+	"github.com/tendermint/go-crypto/keys"
 	abci "github.com/tendermint/tendermint/abci/types"
 )
 
@@ -45,6 +48,11 @@ func (AppModuleBasic) ValidateGenesis(bz json.RawMessage) error {
 		return err
 	}
 	return ValidateGenesis(data)
+}
+
+// register rest routes
+func RegisterRESTRoutes(ctx context.CLIContext, rtr *mux.Router, cdc *codec.Codec, kb keys.Keybase) {
+	rest.RegisterRoutes(ctx, rtr, cdc, StoreKey)
 }
 
 //___________________________
