+++ x/bank/types/codec.go
@@ -1,4 +1,4 @@
-package bank
+package types
 
 import (
 	"github.com/cosmos/cosmos-sdk/codec"
