+++ x/bank/types/msgs.go
@@ -1,4 +1,4 @@
-package bank
+package types
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
