+++ x/bank/types/msgs_test.go
@@ -1,4 +1,4 @@
-package bank
+package types
 
 import (
 	"fmt"
